requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
redis==5.0.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.1
Flask-CORS==4.0.0
Pillow==10.2.0
requests==2.31.0
gunicorn==21.2.0
python-magic==0.4.27
Werkzeug==3.0.1 